+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Separate nav and speedometer UI features into own modules
+- Open game centered on primary monitor
+- Hide cursor
 
 ## [0.0.13] - 2023-10-10